+++ PaymentGateway.Models/Service.cs
@@ -9,9 +9,9 @@ namespace PaymentGateway.Models
     public class Service
     {
         public string Name { get; set; }
-        public double Value { get; set; }
+        public decimal Value { get; set; }
         public string Currency { get; set; }
-        public int Limit { get; set; }
+        public decimal Limit { get; set; }
         public int Id { get; set; }
 
     }
